ChoiceControl, MultiChoiceControl: added $checkAllowedValues [Closes #28]

+++ src/Forms/Controls/ChoiceControl.php
@@ -19,6 +19,9 @@ use Nette;
  */
 abstract class ChoiceControl extends BaseControl
 {
+	/** @var bool */
+	public $checkAllowedValues = TRUE;
+
 	/** @var array */
 	private $items = [];
 
@@ -56,7 +59,7 @@ abstract class ChoiceControl extends BaseControl
 	 */
 	public function setValue($value)
 	{
-		if ($value !== NULL && !array_key_exists((string) $value, $this->items)) {
+		if ($this->checkAllowedValues && $value !== NULL && !array_key_exists((string) $value, $this->items)) {
 			$set = Nette\Utils\Strings::truncate(implode(', ', array_map(function ($s) { return var_export($s, TRUE); }, array_keys($this->items))), 70, '...');
 			throw new Nette\InvalidArgumentException("Value '$value' is out of allowed set [$set] in field '{$this->name}'.");
 		}
