+++ src/Forms/Controls/MultiChoiceControl.php
@@ -19,6 +19,9 @@ use Nette;
  */
 abstract class MultiChoiceControl extends BaseControl
 {
+	/** @var bool */
+	public $checkAllowedValues = TRUE;
+
 	/** @var array */
 	private $items = [];
 
@@ -65,7 +68,7 @@ abstract class MultiChoiceControl extends BaseControl
 			$flip[(string) $value] = TRUE;
 		}
 		$values = array_keys($flip);
-		if ($diff = array_diff($values, array_keys($this->items))) {
+		if ($this->checkAllowedValues && ($diff = array_diff($values, array_keys($this->items)))) {
 			$set = Nette\Utils\Strings::truncate(implode(', ', array_map(function ($s) { return var_export($s, TRUE); }, array_keys($this->items))), 70, '...');
 			$vals = (count($diff) > 1 ? 's' : '') . " '" . implode("', '", $diff) . "'";
 			throw new Nette\InvalidArgumentException("Value$vals are out of allowed set [$set] in field '{$this->name}'.");
